+++ patches/src/main/kotlin/app/revanced/patches/spotify/misc/Fingerprints.kt
@@ -3,7 +3,7 @@ package app.revanced.patches.spotify.misc
 import app.revanced.patcher.fingerprint
 
 internal val accountAttributeFingerprint by fingerprint {
-    custom { _, c -> c.endsWith("internal/AccountAttribute;") }
+    custom { _, classDef -> classDef.endsWith("internal/AccountAttribute;") }
 }
 
 internal val productStateProtoFingerprint by fingerprint {
@@ -15,4 +15,9 @@ internal val productStateProtoFingerprint by fingerprint {
 
 internal val buildQueryParametersFingerprint by fingerprint {
     strings("trackRows", "device_type:tablet")
+}
+
+internal val contextMenuExperimentsFingerprint by fingerprint {
+    parameters("L")
+    strings("remove_ads_upsell_enabled")
 }